+++ .gitignore
@@ -1,3 +1,1 @@
-cell.py
-constants.py
 __pycache__
