+++ highscore.txt
@@ -1,1 +1,1 @@
-35
+0
